app/requirements.txt
django==3.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
djangorestframework==3.13.1
django-phonenumber-field[phonenumbers]==6.1.0
geopy~=2.2.0
requests~=2.27.1
rollbar==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
